+++ src/control_plane/migration/mod.rs
@@ -7,6 +7,7 @@ mod migration_error;
 mod migration_execution_result;
 mod migration_operation_error;
 mod migration_operations;
+mod migration_rollback_plan;
 mod postgres;
 mod run_migration;
 
@@ -16,4 +17,5 @@ pub(crate) use migration_error::MigrationError;
 pub(crate) use migration_execution_result::MigrationExecutionResult;
 pub(crate) use migration_operation_error::MigrationOperationError;
 pub(crate) use migration_operations::{MigrationFuture, MigrationOperations};
+pub(crate) use migration_rollback_plan::MigrationRollbackPlan;
 pub(crate) use run_migration::{confirm_migration, execute_migration, rollback_migration};